Add OGRSpatialReference::exportToCF1() and importFromCF1(), and make netCDF driver use them
