+++ ogr/ogr_srs_api.h
@@ -493,6 +493,9 @@ OGRErr CPL_DLL OSRImportFromMICoordSys(OGRSpatialReferenceH, const char *);
 OGRErr CPL_DLL OSRImportFromERM(OGRSpatialReferenceH, const char *,
                                 const char *, const char *);
 OGRErr CPL_DLL OSRImportFromUrl(OGRSpatialReferenceH, const char *);
+OGRErr CPL_DLL OSRImportFromCF1(OGRSpatialReferenceH,
+                                CSLConstList papszKeyValues,
+                                const char *pszUnits);
 
 OGRErr CPL_DLL CPL_STDCALL OSRExportToWkt(OGRSpatialReferenceH, char **);
 OGRErr CPL_DLL OSRExportToWktEx(OGRSpatialReferenceH, char **ppszResult,
@@ -511,6 +514,9 @@ OGRErr CPL_DLL OSRExportToPanorama(OGRSpatialReferenceH, long *, long *, long *,
                                    long *, double *);
 OGRErr CPL_DLL OSRExportToMICoordSys(OGRSpatialReferenceH, char **);
 OGRErr CPL_DLL OSRExportToERM(OGRSpatialReferenceH, char *, char *, char *);
+OGRErr CPL_DLL OSRExportToCF1(OGRSpatialReferenceH, char **ppszGridMappingName,
+                              char ***ppapszKeyValues, char **ppszUnits,
+                              CSLConstList papszOptions);
 
 OGRErr CPL_DLL OSRMorphToESRI(OGRSpatialReferenceH);
 OGRErr CPL_DLL OSRMorphFromESRI(OGRSpatialReferenceH);
